Feat: Nuevos EndPoints Operaciones
Creo EndPoints de Transfereir, Retirar y VerMovimientos en Operaciones

+++ src/main/java/ar/edu/utn/frbb/tup/service/operaciones/modulos/Transferencia.java
@@ -1,5 +1,6 @@
 package ar.edu.utn.frbb.tup.service.operaciones.modulos;
 
+import ar.edu.utn.frbb.tup.exception.CuentaNoEncontradaException;
 import ar.edu.utn.frbb.tup.model.Cuenta;
 import ar.edu.utn.frbb.tup.model.Movimiento;
 import ar.edu.utn.frbb.tup.persistence.CuentaDao;
@@ -25,54 +26,49 @@ public class Transferencia extends baseOperaciones {
         this.movimientosDao = movimientosDao;
     }
 
-    public void transferencia(Cuenta cuentaOrigen, Cuenta cuentaDestino, double monto){
+    public double transferencia(long cuentaO, long cuenntaD, double monto) throws MismaCuentaException, CuentaEstaDeBajaException, CuentaSinDineroException, CuentaNoEncontradaException {
 
-        try {
-            if (cuentaOrigen.getCVU() == cuentaDestino.getCVU()){ ////Lanzo excepcion cuando la cuenta destino es igual a la origen
-                throw new MismaCuentaException("No se puede transferir a la misma cuenta");
-            }
-            if (!cuentaDestino.getEstado()){ //Lanzo excepcion cuando la cuenta destino esta dada de baja
-                throw new CuentaEstaDeBajaException("La cuenta a transferir esta dada de baja");
-            }
+        Cuenta cuentaOrigen = cuentaDao.findCuenta(cuentaO);
+        Cuenta cuentaDestino = cuentaDao.findCuenta(cuenntaD);
 
-            if (monto > cuentaOrigen.getSaldo()){ //Lanzo excepcion cuando no tiene dinero para trnsferir
-                throw new CuentaSinDineroException("No hay suficiente dinero en la cuenta " + cuentaOrigen.getNombre() + ", su saldo es de $" + cuentaOrigen.getSaldo());
-            }
-
-            System.out.println("----------------------------------------");
-            //Borro la cuentaOrigen ya que va ser modificada
-            cuentaDao.deleteCuenta(cuentaOrigen.getCVU());
-            cuentaDao.deleteCuenta(cuentaDestino.getCVU());
-
-            //Resto el monto a la cuenta origen y sumo a la que se envia
-            cuentaOrigen.setSaldo(cuentaOrigen.getSaldo() - monto);
-            cuentaDestino.setSaldo(cuentaDestino.getSaldo() + monto);
-
-            //Tomo registro de la transferencia en la cuentaOrginen y destino
-            Movimiento movimiento = crearMovimiento(tipoOperacion + cuentaDestino.getNombre(), monto, cuentaOrigen.getCVU());
-            movimientosDao.saveMovimiento(movimiento);
-
-            Movimiento movimientoDestino = crearMovimiento(tipoOperacionDestino + cuentaOrigen.getNombre(), monto, cuentaDestino.getCVU());
-            movimientosDao.saveMovimiento(movimientoDestino);
-
-            //Muestro en pantalla la transferencia
-            System.out.println("Se ha transferido " + monto + " a la cuenta " + cuentaDestino.getNombre());
-
-            //Guardo la cuentaOrigen y cuentaDestino modificada
-            cuentaDao.saveCuenta(cuentaOrigen);
-            cuentaDao.saveCuenta(cuentaDestino);
-
-            System.out.println("----------------------------------------");
-
-        } catch (CuentaSinDineroException | MismaCuentaException | CuentaEstaDeBajaException e) {
-            System.out.println("----------------------------------------");
-            System.out.println(e.getMessage());
-            System.out.println("----------------------------------------");
-        } finally {
-            System.out.println("Enter para seguir");
-            scanner.nextLine();
-            clearScreen();
+        if (cuentaOrigen == null ) {
+            throw new CuentaNoEncontradaException("No se encontro ninguna cuenta con el CVU dado " + cuentaO);
         }
 
+        if (cuentaDestino == null) {
+            throw new CuentaNoEncontradaException("No se encontro ninguna cuenta con el CVU dado " + cuenntaD);
+        }
+
+        if (cuentaOrigen.getCVU() == cuentaDestino.getCVU()) { ////Lanzo excepcion cuando la cuenta destino es igual a la origen
+            throw new MismaCuentaException("No se puede transferir a la misma cuenta");
+        }
+        if (!cuentaDestino.getEstado()) { //Lanzo excepcion cuando la cuenta destino esta dada de baja
+            throw new CuentaEstaDeBajaException("La cuenta a transferir esta dada de baja");
+        }
+
+        if (monto > cuentaOrigen.getSaldo()) { //Lanzo excepcion cuando no tiene dinero para trnsferir
+            throw new CuentaSinDineroException("No hay suficiente dinero en la cuenta " + cuentaOrigen.getNombre() + ", su saldo es de $" + cuentaOrigen.getSaldo());
+        }
+
+        //Borro la cuentaOrigen ya que va ser modificada
+        cuentaDao.deleteCuenta(cuentaOrigen.getCVU());
+        cuentaDao.deleteCuenta(cuentaDestino.getCVU());
+
+        //Resto el monto a la cuenta origen y sumo a la que se envia
+        cuentaOrigen.setSaldo(cuentaOrigen.getSaldo() - monto);
+        cuentaDestino.setSaldo(cuentaDestino.getSaldo() + monto);
+
+        //Tomo registro de la transferencia en la cuentaOrginen y destino
+        Movimiento movimiento = crearMovimiento(tipoOperacion + cuentaDestino.getNombre(), monto, cuentaOrigen.getCVU());
+        movimientosDao.saveMovimiento(movimiento);
+
+        Movimiento movimientoDestino = crearMovimiento(tipoOperacionDestino + cuentaOrigen.getNombre(), monto, cuentaDestino.getCVU());
+        movimientosDao.saveMovimiento(movimientoDestino);
+
+        //Guardo la cuentaOrigen y cuentaDestino modificada
+        cuentaDao.saveCuenta(cuentaOrigen);
+        cuentaDao.saveCuenta(cuentaDestino);
+
+        return cuentaOrigen.getSaldo();
     }
 }
